ml-model-project/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
matplotlib>=3.5.0
pandas>=1.3.0
pyyaml>=6.0
pickle5>=0.0.11